Split new tests into new files
